pipeline/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
langextract==0.4.1
werkzeug==2.3.7
pathlib2==2.3.7
unstructured[all-docs]==0.10.30
